files restructured and sticky headers added
